Move validation logic into separate class
